app/requirements.txt
paddleocr
Pillow
pymupdf
PyPDF2
Shapely
opencv_python
python_Levenshtein
numpy
